backend/requirements.txt
Django>=4.1,<5.0
djangorestframework>=3.14.0
djongo>=1.3.6
pymongo>=3.12.3,<4.0
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.3.0
cryptography>=41.0.0
pyotp>=2.9.0
drf-yasg>=1.21.7
django-cors-headers>=4.2.0
gunicorn>=21.2.0
requests>=2.31.0
user-agents>=2.2.0
django-filter>=23.2
